+++ frontend/src/components/BpmnModeler/properties-provider/SequenceFlowPropertiesPanel.d.ts
@@ -1,0 +1,15 @@
+import React from 'react';
+
+interface SequenceFlowPropertiesPanelProps {
+  element: unknown;
+  modeler: unknown;
+  variables?: Array<{
+    name: string;
+    type?: string;
+    [key: string]: unknown;
+  }>;
+}
+
+declare const SequenceFlowPropertiesPanel: React.FC<SequenceFlowPropertiesPanelProps>;
+
+export default SequenceFlowPropertiesPanel;